requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.11.0
pydantic-settings==2.3.3
boto3==1.34.122
httpx==0.27.0
pdfminer.six==20231228
